// packages/fe-merchant-pc/src/components/PaperExamine/editTmp/interface.ts
export interface ExamineDetailType {
    authenticateCode: string
    authenticateTitle: string
    chaosOptionsState: number
    chaosOrderState: number
    checkQuestionType: boolean
    checkRandomQuestionNum: boolean
    code: string
    composition: string
    customContent: Record<string, any>
    haveUnificationScore: boolean
    numContinuousState: number
    organizationCode: string
    precautions: string
    qualifiedProp: number
    quoteNumStatus: number
    questionCitedLimit: number
    questionConfigList: {
        needNumber: number
        questionType: string
        score: number
    }[]
    questionStructure: string
    questionTotal: number
    questionTypeLeast: number
    randomQuestionNum: number
    randomQuestionState: number
    receiptEndTime: string
    receiptStartTime: string
    savedQuestionTotal: number
    scoreTotal: number
    scoreType: string
    sid: number
    suggestedTime: number
    tags: string
    templateCode: string
    templateTitle: string
    title: string
    trueState: number
    unificationScore: number
    userCode: string
    advancedConfig: string
    totalScore: string
    groupCount: number // 重复试题组数
    taskCode: string // 轮询重复试题的任务code
}

/**
 * 鉴定点详情
 */
export interface AuthDetailType {
    detailList: {
        actualGravity: string
        pointList: {
            difficulty: string
            distinction: string
            important: string
            point: string
            pointCode: string
            questionTotal: number
            questionType: string
            // 自定义字段
            rowSpan: number
        }[]
        range: string
        rangeCode: string
        requiredGravity: string
    }[]
    statistics: {
        importantX: number
        importantY: number
        importantZ: number
    }
}
export interface AuthDetailListType {
    actualGravity: string
    difficulty: string
    distinction: string
    important: string
    point: string
    pointCode: string
    questionTotal: number
    questionType: string
    // 自定义字段
    rowSpan: number
    range: string
    rangeCode: string
    requiredGravity: string
}
export interface PublishType {
    success: boolean
    failReasons: string[]
}